Fixes sessions to use ScriptBlocks, stopping crashes and improved performance

+++ src/Private/Sessions.ps1
@@ -90,7 +90,7 @@ function Get-PodeSession
 
         # get the header from the request
         $value = Get-PodeHeader -Name $Session.Name -Secret $secret
-        if (Test-IsEmpty $value) {
+        if ([string]::IsNullOrWhiteSpace($value)) {
             return $null
         }
     }
@@ -104,7 +104,7 @@ function Get-PodeSession
 
         # get the cookie from the request
         $cookie = Get-PodeCookie -Name $Session.Name -Secret $secret
-        if (Test-IsEmpty $cookie) {
+        if ([string]::IsNullOrWhiteSpace($cookie)) {
             return $null
         }
 
@@ -156,7 +156,10 @@ function Set-PodeSessionDataHash
         $Session
     )
 
-    $Session.Data = (Protect-PodeValue -Value $Session.Data -Default @{})
+    if (($null -eq $Session.Data) -or ($Session.Data.Count -eq 0)) {
+        $Session.Data = @{}
+    }
+
     $Session.DataHash = (Invoke-PodeSHA256Hash -Value ($Session.Data | ConvertTo-Json -Depth 10 -Compress))
 }
 
@@ -169,11 +172,14 @@ function Test-PodeSessionDataHash
         $Session
     )
 
-    if (Test-IsEmpty $Session.DataHash) {
+    if ([string]::IsNullOrWhiteSpace($Session.DataHash)) {
         return $false
     }
 
-    $Session.Data = (Protect-PodeValue -Value $Session.Data -Default @{})
+    if (($null -eq $Session.Data) -or ($Session.Data.Count -eq 0)) {
+        $Session.Data = @{}
+    }
+
     $hash = (Invoke-PodeSHA256Hash -Value ($Session.Data | ConvertTo-Json -Depth 10 -Compress))
     return ($Session.DataHash -eq $hash)
 }
@@ -191,7 +197,11 @@ function Get-PodeSessionExpiry
         return [DateTime]::MinValue
     }
 
-    $expiry = (Resolve-PodeValue -Check ([bool]$Session.Properties.Extend) -TrueValue ([DateTime]::UtcNow) -FalseValue $Session.Properties.TimeStamp)
+    $expiry = [DateTime]::UtcNow
+    if (!([bool]$Session.Properties.Extend)) {
+        $expiry = $Session.Properties.TimeStamp
+    }
+
     $expiry = $expiry.AddSeconds($Session.Properties.Duration)
     return $expiry
 }
@@ -223,7 +233,12 @@ function Set-PodeSessionHelpers
         $expiry = (Get-PodeSessionExpiry -Session $session)
 
         # save session data to store
-        $PodeContext.Server.Sessions.Store.Set($session.Id, $session.Data, $expiry)
+        if ($PodeContext.Server.Sessions.Store.Set -is [psscriptmethod]) {
+            $PodeContext.Server.Sessions.Store.Set($session.Id, $session.Data, $expiry)
+        }
+        else {
+            Invoke-PodeScriptBlock -ScriptBlock $PodeContext.Server.Sessions.Store.Set -Arguments @($session.Id, $session.Data, $expiry) -Splat
+        }
 
         # update session's data hash
         Set-PodeSessionDataHash -Session $session
@@ -234,7 +249,12 @@ function Set-PodeSessionHelpers
         param($session)
 
         # remove data from store
-        $PodeContext.Server.Sessions.Store.Delete($session.Id)
+        if ($PodeContext.Server.Sessions.Store.Delete -is [psscriptmethod]) {
+            $PodeContext.Server.Sessions.Store.Delete($session.Id)
+        }
+        else {
+            Invoke-PodeScriptBlock -ScriptBlock $PodeContext.Server.Sessions.Store.Delete -Arguments $session.Id
+        }
 
         # clear session
         $session.Clear()
@@ -249,20 +269,20 @@ function Get-PodeSessionInMemStore
     $store | Add-Member -MemberType NoteProperty -Name Memory -Value @{}
 
     # delete a sessionId and data
-    $store | Add-Member -MemberType ScriptMethod -Name Delete -Value {
+    $store | Add-Member -MemberType NoteProperty -Name Delete -Value {
         param($sessionId)
-        $this.Memory.Remove($sessionId) | Out-Null
+        $PodeContext.Server.Sessions.Store.Memory.Remove($sessionId) | Out-Null
     }
 
     # get a sessionId's data
-    $store | Add-Member -MemberType ScriptMethod -Name Get -Value {
+    $store | Add-Member -MemberType NoteProperty -Name Get -Value {
         param($sessionId)
 
-        $s = $this.Memory[$sessionId]
+        $s = $PodeContext.Server.Sessions.Store.Memory[$sessionId]
 
         # if expire, remove
         if (($null -ne $s) -and ($s.Expiry -lt [DateTime]::UtcNow)) {
-            $this.Memory.Remove($sessionId) | Out-Null
+            $PodeContext.Server.Sessions.Store.Memory.Remove($sessionId) | Out-Null
             return $null
         }
 
@@ -270,10 +290,10 @@ function Get-PodeSessionInMemStore
     }
 
     # update/insert a sessionId and data
-    $store | Add-Member -MemberType ScriptMethod -Name Set -Value {
+    $store | Add-Member -MemberType NoteProperty -Name Set -Value {
         param($sessionId, $data, $expiry)
 
-        $this.Memory[$sessionId] = @{
+        $PodeContext.Server.Sessions.Store.Memory[$sessionId] = @{
             Data = $data
             Expiry = $expiry
         }
@@ -308,7 +328,23 @@ function Set-PodeSessionInMemClearDown
 
 function Test-PodeSessionsConfigured
 {
-    return (!(Test-IsEmpty $PodeContext.Server.Sessions))
+    return (($null -ne $PodeContext.Server.Sessions) -and ($PodeContext.Server.Sessions.Count -gt 0))
+}
+
+function Get-PodeSessionData
+{
+    param(
+        [Parameter()]
+        [string]
+        $SessionId
+    )
+
+    if ($PodeContext.Server.Sessions.Store.Get -is [psscriptmethod]) {
+        return $PodeContext.Server.Sessions.Store.Get($e.Session.Id)
+    }
+    else {
+        return (Invoke-PodeScriptBlock -ScriptBlock $PodeContext.Server.Sessions.Store.Get -Arguments $SessionId -Return)
+    }
 }
 
 function Get-PodeSessionMiddleware
@@ -333,7 +369,7 @@ function Get-PodeSessionMiddleware
             }
 
             # get the session's data
-            elseif ($null -ne ($data = $PodeContext.Server.Sessions.Store.Get($e.Session.Id))) {
+            elseif ($null -ne ($data = (Get-PodeSessionData -SessionId $e.Session.Id))) {
                 $e.Session.Data = $data
                 Set-PodeSessionDataHash -Session $e.Session
             }
